Files restructing
