+++ internal/services/weather/types.go
