+++ internal/services/weather/yandex/defaults.go
