+++ internal/services/weather/yandex/types.go
